Display refactoring
