+++ Controllers/GameController.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using UnityEngine;
-using UnityEngine.AI;
-using UnityEngine.SceneManagement;
-using UnityEngine.UI;
 
 
 namespace Labyrinth
@@ -10,21 +7,14 @@ namespace Labyrinth
     public sealed class GameController : MonoBehaviour, IDisposable
     {
         #region Field
-        
+
         public PlayerType PlayerType = PlayerType.Ball;
         private ListExecuteObject _interactiveObject;
-        private DisplayEndGame _displayEndGame;
-        private DisplayBonuses _displayBonuses;
-        private DisplayWinGame _displayWinGame;
-        private DisplaySpeed _displaySpeed;
-        private Button _restartButton;
+        private ViewInitializer _viewInitializer;
         private CameraController _cameraController;
         private InputController _inputController;
-        private Reference _reference;
         private HoleBonus _holeBonus;
         private PlayerBase _player;
-        private int _countBonuses;
-        private int _winBonusRemained;
 
         #endregion
 
@@ -34,57 +24,50 @@ namespace Labyrinth
         private void Awake()
         {
             _interactiveObject = new ListExecuteObject();
-            
+
             var reference = new Reference();
             var bonusReference = new BonusReference();
-            
+
             _player = null;
             if (PlayerType == PlayerType.Ball)
             {
                 _player = reference.PlayerBall;
             }
-            
+
             _holeBonus = bonusReference.HoleBonus;
             _interactiveObject.AddExecuteObject(_holeBonus);
-            
+
             _cameraController = new CameraController(_player.transform, reference.MainCamera.transform);
             _interactiveObject.AddExecuteObject(_cameraController);
-            
+
             _inputController = new InputController(_player, _interactiveObject);
             _interactiveObject.AddExecuteObject(_inputController);
-            
-            _displayBonuses = new DisplayBonuses(reference.Bonus);
-            _displayEndGame = new DisplayEndGame(reference.EndGame);
-            _displayWinGame = new DisplayWinGame(reference.WinGame);
-            _displaySpeed = new DisplaySpeed(reference.SpeedDisplay);
-            
-            _player.ShowSpeedAction += ShowNewSpeed;
-            
-            _restartButton = reference.RestartButton;
-            
+
+            _viewInitializer = new ViewInitializer();
+            _viewInitializer.InitializeDisplay();
+            _player.ShowSpeedAction += _viewInitializer.ShowNewSpeed;
+
             if (Application.platform == RuntimePlatform.WindowsEditor)
             {
                 _inputController = new InputController(_player, _interactiveObject);
                 _interactiveObject.AddExecuteObject(_inputController);
             }
-            
+
             foreach (var o in _interactiveObject)
             {
                 if (o is HoleBonus holeBonus)
                 {
-                    holeBonus.OnCaughtPlayerChange += CaughtPlayer;
-                    holeBonus.OnCaughtPlayerChange += _displayEndGame.ShowLoseGameLabel;
+                    holeBonus.OnCaughtPlayerChange += _viewInitializer.CaughtPlayer;
                 }
-                
+
                 if (o is WinBonus winBonus)
                 {
-                    winBonus.OnPointChange += AddBonus;
-                    _winBonusRemained++;
+                    winBonus.OnPointChange += _viewInitializer.AddBonus;
+                    _viewInitializer.WinBonusRemained++;
                 }
             }
-            _restartButton.onClick.AddListener(RestartGame);
-            _restartButton.gameObject.SetActive(false);
-            ShowNewSpeed(_player.Speed);
+
+            _viewInitializer.ShowNewSpeed(_player.Speed);
         }
 
         private void Update()
@@ -96,6 +79,7 @@ namespace Labyrinth
                 {
                     continue;
                 }
+
                 interactiveObject.Execute();
             }
         }
@@ -104,43 +88,6 @@ namespace Labyrinth
 
 
         #region Methods
-        
-        private void CaughtPlayer(string value, Color args)
-        {
-            _restartButton.gameObject.SetActive(true);
-            Time.timeScale = 0.0f;
-        }
-        
-        private void AddBonus(int value)
-        {
-            _countBonuses += value;
-            _displayBonuses.Display(_countBonuses);
-            _winBonusRemained--;
-            if (_winBonusRemained == 0)
-            {
-                СongratulateVictory();
-                _restartButton.gameObject.SetActive(true);
-            }
-        }
-
-        private void RestartGame()
-        {
-            Dispose();
-            SceneManager.LoadScene(sceneBuildIndex: 0); 
-            Time.timeScale = 1.0f;
-        }
-
-        private void СongratulateVictory()
-        {
-            Dispose();
-            _displayWinGame.WinGame();
-            Time.timeScale = 0.0f;
-        }
-
-        private void ShowNewSpeed(float newSpeed)
-        {
-            _displaySpeed.ShowSpeed(newSpeed);
-        }
 
         public void Dispose()
         {
@@ -148,18 +95,18 @@ namespace Labyrinth
             {
                 if (o is HoleBonus holeBonus)
                 {
-                    holeBonus.OnCaughtPlayerChange -= CaughtPlayer;
-                    holeBonus.OnCaughtPlayerChange -= _displayEndGame.ShowLoseGameLabel;
+                    holeBonus.OnCaughtPlayerChange -= _viewInitializer.CaughtPlayer;
                 }
-            
+
                 if (o is WinBonus winBonus)
                 {
-                    winBonus.OnPointChange -= AddBonus;
+                    winBonus.OnPointChange -= _viewInitializer.AddBonus;
                 }
             }
-            _player.ShowSpeedAction -= ShowNewSpeed;
+
+            _player.ShowSpeedAction -= _viewInitializer.ShowNewSpeed;
         }
-        
+
         #endregion
     }
 }
