+++ Interface/IView.cs
@@ -1,0 +1,9 @@
+﻿
+namespace Labyrinth
+
+{
+    public interface IView
+    {
+        
+    }
+}
