+++ Model/PlayerBase.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using static Labyrinth.TimeScaleForBonus;
@@ -14,12 +13,14 @@ namespace Labyrinth
         public Dictionary<string, Action> SpeedBonusActions;
         public event Action<float> ShowSpeedAction;
 
-        public float Speed = 3.0f;
+        private float _speed = 3.0f;
         public abstract void Move(float x, float y, float z);
 
         private float _speedChanger = 2.0f;
         private float _baseSpeed;
         private int _interval = 10;
+
+        public float Speed => _speed;
 
         #endregion
 
@@ -51,35 +52,23 @@ namespace Labyrinth
 
         private void SpeedDown()
         {
-            Speed /= _speedChanger;
-            ShowSpeedAction?.Invoke(Speed);
+            _speed /= _speedChanger;
+           ShowSpeedAction?.Invoke(Speed);
             SpeedBase();
         }
 
         private void SpeedUp()
         {
-            Speed *= _speedChanger;
+            _speed *= _speedChanger;
             ShowSpeedAction?.Invoke(Speed);
             SpeedBase();
         }
 
         private void ReturnBaseSpeed()
         {
-            Speed = _baseSpeed;
+            _speed = _baseSpeed;
             ShowSpeedAction?.Invoke(Speed);
         }
-        //
-        // public void Dead()
-        // {
-        //     Destroy(gameObject);
-        // }
-
-        // IEnumerator TimeBonus(int _interval)
-        // {
-        //     yield return new WaitForSeconds(_interval);
-        //
-        //     ReturnBaseSpeed();
-        // }
 
         #endregion
     }
